requirements.txt
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
Automat==24.8.1
awscli==1.22.34
blinker==1.4
botocore==1.23.34
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
click==8.1.7
colorama==0.4.4
command-not-found==0.3
constantly==23.10.4
cryptography==43.0.1
cssselect==1.2.0
dbus-python==1.2.18
defusedxml==0.7.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.7.0
docutils==0.17.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.2
fastapi-cli==0.0.5
filelock==3.16.0
gyp==0.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.1
jeepney==0.7.1
Jinja2==3.1.4
jmespath==0.10.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==24.1
parsel==1.9.1
Pillow==9.0.1
Protego==0.3.1
psutil==6.0.0
pyasn1==0.4.8
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyDispatcher==2.0.7
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyOpenSSL==24.2.1
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==5.4.1
queuelib==1.7.0
requests==2.25.1
requests-file==2.1.0
rich==13.9.2
roman==3.3
rsa==4.8
s3transfer==0.5.0
Scrapy==2.11.2
SecretStorage==3.3.1
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.39.2
systemd-python==234
tldextract==5.1.2
tomli==2.0.1
Twisted==24.7.0
typer==0.12.5
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.31.1
uvloop==0.20.0
w3lib==2.2.1
wadllib==1.3.6
watchfiles==0.24.0
websockets==13.1
zipp==1.0.0
zope.interface==7.0.3
